권한 요청 추가 이미지 가져오고 적용 채팅 Activty 추가
