+++ app/src/main/AndroidManifest.xml
@@ -22,6 +22,7 @@
         <activity android:name=".profile.ProfileActivity" />
         <activity android:name=".profile.PhotoActivity" />
         <activity android:name=".profile.ProfileSettingActivity" />
+        <activity android:name=".chatting.ChattingActivity" />
         <activity android:name="com.theartofdev.edmodo.cropper.CropImageActivity"
             android:theme="@style/Base.Theme.AppCompat"/>
     </application>
