+++ app/src/main/java/kr/pe/dreamer/startalk/chatting/ChattingActivity.java
@@ -1,0 +1,20 @@
+package kr.pe.dreamer.startalk.chatting;
+
+import android.os.Bundle;
+import android.os.PersistableBundle;
+import android.support.v7.app.AppCompatActivity;
+
+import kr.pe.dreamer.startalk.R;
+
+/**
+ * Created by rhdlr on 2017-02-26.
+ */
+
+public class ChattingActivity extends AppCompatActivity {
+    @Override
+    public void onCreate(Bundle savedInstanceState, PersistableBundle persistentState) {
+        super.onCreate(savedInstanceState, persistentState);
+        setContentView(R.layout.activity_chatroom);
+
+    }
+}
